APIGOV-22058 - split cache manager and add crd and ard methods
