+++ pkg/agent/cache/manager.go
@@ -2,7 +2,6 @@ package cache
 
 import (
 	"encoding/json"
-	"fmt"
 	"sync"
 
 	defs "github.com/Axway/agent-sdk/pkg/apic/definitions"
@@ -15,10 +14,7 @@ import (
 	"github.com/Axway/agent-sdk/pkg/util/log"
 )
 
-const (
-	defaultCacheStoragePath = "./data/cache"
-	accessControlList       = "AccessControlListKey"
-)
+const defaultCacheStoragePath = "./data/cache"
 
 // Manager - interface to manage agent resource
 type Manager interface {
@@ -61,6 +57,16 @@ type Manager interface {
 	SetAccessControlList(acl *v1.ResourceInstance)
 	GetAccessControlList() *v1.ResourceInstance
 
+	// AccessRequestDefintion cache related methods
+	AddAccessRequestDefinition(resource *v1.ResourceInstance)
+	GetAccessRequestDefinitionByName(name string) *v1.ResourceInstance
+	DeleteAccessRequestDefinitionByName(name string) error
+
+	// CredentialRequestDefintion cache related methods
+	AddCredentialRequestDefinition(resource *v1.ResourceInstance)
+	GetCredentialRequestDefinitionByName(name string) *v1.ResourceInstance
+	DeleteCredentialRequestDefinitionByName(name string) error
+
 	// Watch Sequence cache related methods
 	AddSequence(watchTopicName string, sequenceID int64)
 	GetSequence(watchTopicName string) int64
@@ -79,6 +85,8 @@ type cacheManager struct {
 	cacheLock               sync.Mutex
 	persistedCache          cache.Cache
 	teams                   cache.Cache
+	ardMap                  cache.Cache
+	crdMap                  cache.Cache
 	cacheFilename           string
 	hasLoadedPersistedCache bool
 	isCacheUpdated          bool
@@ -92,6 +100,8 @@ func NewAgentCacheManager(cfg config.CentralConfig, persistCache bool) Manager {
 		categoryMap:    cache.New(),
 		sequenceCache:  cache.New(),
 		teams:          cache.New(),
+		ardMap:         cache.New(),
+		crdMap:         cache.New(),
 		isCacheUpdated: false,
 	}
 
@@ -201,313 +211,6 @@ func (c *cacheManager) SaveCache() {
 		c.persistedCache.Save(c.cacheFilename)
 		c.setCacheUpdated(false)
 	}
-}
-
-// API service cache management
-
-// AddAPIService - add/update APIService resource in cache
-func (c *cacheManager) AddAPIService(svc *v1.ResourceInstance) error {
-	apiID, err := util.GetAgentDetailsValue(svc, defs.AttrExternalAPIID)
-	if err != nil {
-		return fmt.Errorf("failed to get external API ID from APIService resource: %s", err)
-	}
-	if apiID != "" {
-		defer c.setCacheUpdated(true)
-		apiName, _ := util.GetAgentDetailsValue(svc, defs.AttrExternalAPIName)
-		primaryKey, _ := util.GetAgentDetailsValue(svc, defs.AttrExternalAPIPrimaryKey)
-		if primaryKey != "" {
-			// Verify secondary key and validate if we need to remove it from the apiMap (cache)
-			if _, err := c.apiMap.Get(apiID); err != nil {
-				c.apiMap.Delete(apiID)
-			}
-
-			c.apiMap.SetWithSecondaryKey(primaryKey, apiID, svc)
-			c.apiMap.SetSecondaryKey(primaryKey, apiName)
-		} else {
-			c.apiMap.SetWithSecondaryKey(apiID, apiName, svc)
-		}
-		log.Tracef("added api name: %s, id %s to API cache", apiName, apiID)
-	}
-
-	return nil
-}
-
-// GetAPIServiceCache - returns the APIService cache
-func (c *cacheManager) GetAPIServiceCache() cache.Cache {
-	return c.apiMap
-}
-
-// GetAPIServiceKeys - returns keys for APIService cache
-func (c *cacheManager) GetAPIServiceKeys() []string {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	return c.apiMap.GetKeys()
-}
-
-// GetAPIServiceWithAPIID - returns resource from APIService cache based on externalAPIID attribute
-func (c *cacheManager) GetAPIServiceWithAPIID(apiID string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	api, _ := c.apiMap.Get(apiID)
-	if api == nil {
-		api, _ = c.apiMap.GetBySecondaryKey(apiID)
-	}
-
-	if api != nil {
-		apiSvc, ok := api.(*v1.ResourceInstance)
-		if ok {
-			return apiSvc
-		}
-	}
-	return nil
-}
-
-// GetAPIServiceWithPrimaryKey - returns resource from APIService cache based on externalAPIPrimaryKey attribute
-func (c *cacheManager) GetAPIServiceWithPrimaryKey(primaryKey string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	api, _ := c.apiMap.Get(primaryKey)
-	if api != nil {
-		apiSvc, ok := api.(*v1.ResourceInstance)
-		if ok {
-			return apiSvc
-		}
-	}
-	return nil
-}
-
-// GetAPIServiceWithName - returns resource from APIService cache based on externalAPIName attribute
-func (c *cacheManager) GetAPIServiceWithName(apiName string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	api, _ := c.apiMap.GetBySecondaryKey(apiName)
-	if api != nil {
-		apiSvc, ok := api.(*v1.ResourceInstance)
-		if ok {
-			return apiSvc
-		}
-	}
-	return nil
-}
-
-// GetTeamsIDsInAPIServices - returns the array of team IDs that have services
-func (c *cacheManager) GetTeamsIDsInAPIServices() []string {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	teamNameMap := make(map[string]struct{})
-	teamIDs := make([]string, 0)
-	for _, key := range c.apiMap.GetKeys() {
-		api, _ := c.apiMap.Get(key)
-		if apiSvc, ok := api.(*v1.ResourceInstance); ok {
-			if apiSvc.Owner != nil && apiSvc.Owner.Type == v1.TeamOwner {
-				if _, found := teamNameMap[apiSvc.Owner.ID]; found {
-					continue
-				}
-				teamNameMap[apiSvc.Owner.ID] = struct{}{}
-				teamIDs = append(teamIDs, apiSvc.Owner.ID)
-			}
-		}
-	}
-
-	return teamIDs
-}
-
-// DeleteAPIService - remove APIService resource from cache based on externalAPIID or externalAPIPrimaryKey
-func (c *cacheManager) DeleteAPIService(key string) error {
-	defer c.setCacheUpdated(true)
-
-	err := c.apiMap.Delete(key)
-	if err != nil {
-		err = c.apiMap.DeleteBySecondaryKey(key)
-	}
-	return err
-}
-
-// API service instance management
-
-// AddAPIServiceInstance -  add/update APIServiceInstance resource in cache
-func (c *cacheManager) AddAPIServiceInstance(resource *v1.ResourceInstance) {
-	defer c.setCacheUpdated(true)
-
-	c.instanceMap.Set(resource.Metadata.ID, resource)
-}
-
-// GetAPIServiceInstanceKeys - returns keys for APIServiceInstance cache
-func (c *cacheManager) GetAPIServiceInstanceKeys() []string {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	return c.instanceMap.GetKeys()
-}
-
-// GetAPIServiceInstanceByID - returns resource from APIServiceInstance cache based on instance ID
-func (c *cacheManager) GetAPIServiceInstanceByID(id string) (*v1.ResourceInstance, error) {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	item, err := c.instanceMap.Get(id)
-	if item != nil {
-		instance, ok := item.(*v1.ResourceInstance)
-		if ok {
-			return instance, nil
-		}
-	}
-	return nil, err
-}
-
-// DeleteAPIServiceInstance - remove APIServiceInstance resource from cache based on instance ID
-func (c *cacheManager) DeleteAPIServiceInstance(id string) error {
-	defer c.setCacheUpdated(true)
-
-	return c.instanceMap.Delete(id)
-}
-
-// DeleteAllAPIServiceInstance - remove all APIServiceInstance resource from cache
-func (c *cacheManager) DeleteAllAPIServiceInstance() {
-	defer c.setCacheUpdated(true)
-
-	c.instanceMap.Flush()
-}
-
-// Category cache management
-
-// AddCategory - add/update Category resource in cache
-func (c *cacheManager) AddCategory(resource *v1.ResourceInstance) {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-	defer c.setCacheUpdated(true)
-
-	c.categoryMap.SetWithSecondaryKey(resource.Name, fmt.Sprintf("title-%s", resource.Title), resource)
-}
-
-// GetCategoryCache - returns the Category cache
-func (c *cacheManager) GetCategoryCache() cache.Cache {
-	return c.categoryMap
-}
-
-// GetCategoryKeys - returns keys for Category cache
-func (c *cacheManager) GetCategoryKeys() []string {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	return c.categoryMap.GetKeys()
-}
-
-// GetCategory - returns resource from Category cache based on name
-func (c *cacheManager) GetCategory(name string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	category, _ := c.categoryMap.Get(name)
-	if category != nil {
-		ri, ok := category.(*v1.ResourceInstance)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-// GetCategoryWithTitle - returns resource from Category cache based on title
-func (c *cacheManager) GetCategoryWithTitle(title string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	category, _ := c.categoryMap.GetBySecondaryKey(fmt.Sprintf("title-%s", title))
-	if category != nil {
-		ri, ok := category.(*v1.ResourceInstance)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-// DeleteCategory - remove Category resource from cache based on name
-func (c *cacheManager) DeleteCategory(name string) error {
-	defer c.setCacheUpdated(true)
-
-	return c.categoryMap.Delete(name)
-}
-
-// GetTeamCache - returns the team cache
-func (c *cacheManager) GetTeamCache() cache.Cache {
-	return c.teams
-}
-
-// AddTeam saves a team to the cache
-func (c *cacheManager) AddTeam(team *defs.PlatformTeam) {
-	defer c.setCacheUpdated(true)
-	c.teams.SetWithSecondaryKey(team.Name, team.ID, *team)
-}
-
-// GetTeamByName gets a team by name
-func (c *cacheManager) GetTeamByName(name string) *defs.PlatformTeam {
-	item, err := c.teams.Get(name)
-	if err != nil {
-		return nil
-	}
-	team, ok := item.(defs.PlatformTeam)
-	if !ok {
-		return nil
-	}
-	return &team
-}
-
-// GetDefaultTeam gets the default team
-func (c *cacheManager) GetDefaultTeam() *defs.PlatformTeam {
-	names := c.teams.GetKeys()
-
-	var defaultTeam defs.PlatformTeam
-	for _, name := range names {
-		item, _ := c.teams.Get(name)
-		team, ok := item.(defs.PlatformTeam)
-		if !ok {
-			continue
-		}
-
-		if team.Default {
-			defaultTeam = team
-			break
-		}
-
-		continue
-	}
-
-	return &defaultTeam
-}
-
-// GetTeamByID gets a team by id
-func (c *cacheManager) GetTeamByID(id string) *defs.PlatformTeam {
-	item, err := c.teams.GetBySecondaryKey(id)
-	if err != nil {
-		return nil
-	}
-	team, ok := item.(defs.PlatformTeam)
-	if !ok {
-		return nil
-	}
-	return &team
-}
-
-// SetAccessControlList saves the Access Control List to the cache
-func (c *cacheManager) SetAccessControlList(acl *v1.ResourceInstance) {
-	defer c.setCacheUpdated(true)
-	c.teams.Set(accessControlList, acl)
-}
-
-// GetAccessControlList gets the Access Control List from the cache
-func (c *cacheManager) GetAccessControlList() *v1.ResourceInstance {
-	item, err := c.teams.Get(accessControlList)
-	if err != nil {
-		return nil
-	}
-	return item.(*v1.ResourceInstance)
 }
 
 // Watch Sequence cache
